definiton of boundary conditions seperated from calculation functions
